+++ tests/testthat/test_Graph.R
@@ -436,3 +436,18 @@ test_that("dot output", {
     "6 [label=\"OUTPUT",
     "nop_output\",fontsize=24]"), out[-c(1L, 15L)])
 })
+
+test_that("help() call", {
+  if (identical(help, utils::help)) {  # different behaviour if pkgload / devtools are doing help vs. vanilla R help()
+    # c() to drop attributes
+    expect_equal(
+      c(help("Graph", package = "mlr3pipelines")),
+      c((po("scale") %>>% po("nop"))$help())
+    )
+  } else {
+    expect_equal(
+      help("Graph", package = "mlr3pipelines"),
+      (po("scale") %>>% po("nop"))$help()
+    )
+  }
+})